keep root id alive
When L3_decRef is called and there is only one user of the linked rootid. The rootid of the linked file is release.
This leads to a BAD ROOT error later as we do not reopen the root node.
This patch should help with multiple links to the same file but the impact on #91 is unknown.